lambda/image-black-white-opencv/image/requirements.txt
Pillow==8.4.0
opencv-python-headless==4.5.3.56